surgicalprep-backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation and settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP client (for webhooks, etc.)
httpx==0.26.0

# Stripe (for subscriptions)
stripe==8.2.0

# Cloudflare R2 Storage (S3-compatible)
boto3==1.34.0

# Utilities
python-dotenv==1.0.1

# Development/Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
